Update patch set 2
Patch Set 2:

WDYT about this check? Do you think it might be useful to add it, I sometimes forget to add the Copyright header to new files.

The check is not optimal since it is checking all the files and not only the one just commit, but it might be a starting point.

Patch-set: 2
Work-in-progress: false